Updated experiment to output magnitude differently for graphing spectrogram

+++ Source/Experiment/GPExperiment.h
@@ -114,7 +114,7 @@ private:
     // GRAPH HELPERS
     void fillFrequencyAxisBuffer(unsigned fftSize, double sr, float* buffer);
     String floatBuffersToGraphText(String options, String xlab, String ylab, bool indexAsX, unsigned n, const float* x, const float* y, const float* z);
-    String doubleBuffersToGraphText(String options, String xlab, String ylab, String zlab, bool indexAsX, unsigned n, const float* x, const double* y, const double* z);
+    String doubleBuffersToGraphText(String options, String xlab, String ylab, String zlab, bool indexAsX, unsigned n, const double* x, const double* y, const double* z);
 };
 
 #endif
